requirements.txt
Flask
twilio
requests
pydub
SpeechRecognition
python-dotenv
openai==0.28
transformers==4.41.1
sentencepiece
tensorflow==2.16.1
tf_keras==2.16.0
